:bath: and :wrench: menus
