+++ backend/main.development.js
@@ -1,4 +1,4 @@
-import {app, BrowserWindow, Menu, shell} from 'electron';
+import {app, BrowserWindow} from 'electron';
 import {contains} from 'ramda';
 import {Logger} from './logger';
 import {SequelizeManager, OPTIONS} from './sequelizeManager';
@@ -6,12 +6,10 @@ import {ipcMessageReceive,
         serverMessageReceive,
         channel} from './messageHandler';
 import {setupHTTP, setupHTTPS, findSelfSignedCert} from './setupServers';
-
+import {setupMenus} from './menus';
 
 const ipcMain = require('electron').ipcMain;
 
-let menu;
-let template;
 let mainWindow = null;
 
 if (process.env.NODE_ENV === 'development') {
@@ -69,20 +67,16 @@ app.on('ready', () => {
         });
 
         if (canSetupHTTPS) {
-
             const hasSelfSignedCert = findSelfSignedCert();
-
             mainWindow.webContents.send(channel, {
                     hasSelfSignedCert: hasSelfSignedCert
             });
-
             if (hasSelfSignedCert) {
                 setupHTTPS({
                     sequelizeManager, serverMessageReceive,
                     mainWindow, OPTIONS
                 });
             }
-
         }
 
     });
@@ -95,170 +89,6 @@ app.on('ready', () => {
         mainWindow.openDevTools();
     }
 
-    const help = {
-        label: 'Help',
-        submenu: [{
-            label: 'Documentation',
-            click() {
-                shell.openExternal('https://help.plot.ly/database-connectors/');
-            }
-        }, {
-            label: 'View Repository',
-            click() {
-                shell.openExternal('https://github.com/plotly/' +
-                    'plotly-database-connector');
-            }
-        }, {
-            label: 'Search Issues',
-            click() {
-                shell.openExternal('https://github.com/plotly/' +
-                    'plotly-database-connector/issues');
-            }
-        }]
-    };
+    setupMenus(app, mainWindow);
 
-    if (process.platform === 'darwin') {
-        template = [{
-            label: 'Plotly Database Connector',
-            submenu: [{
-                label: 'About Plotly Database Connector',
-                selector: 'orderFrontStandardAboutPanel:'
-            }, {
-                type: 'separator'
-            }, {
-                label: 'Services',
-                submenu: []
-            }, {
-                type: 'separator'
-            }, {
-                label: 'Hide Plotly Database Connector',
-                accelerator: 'Command+H',
-                selector: 'hide:'
-            }, {
-                label: 'Hide Others',
-                accelerator: 'Command+Shift+H',
-                selector: 'hideOtherApplications:'
-            }, {
-                label: 'Show All',
-                selector: 'unhideAllApplications:'
-            }, {
-                type: 'separator'
-            }, {
-                label: 'Quit',
-                accelerator: 'Command+Q',
-                click() {
-                    app.quit();
-                }
-            }]
-        }, {
-            label: 'Edit',
-            submenu: [{
-                type: 'separator'
-            }, {
-                label: 'Cut',
-                accelerator: 'Command+X',
-                selector: 'cut:'
-            }, {
-                label: 'Copy',
-                accelerator: 'Command+C',
-                selector: 'copy:'
-            }, {
-                label: 'Paste',
-                accelerator: 'Command+V',
-                selector: 'paste:'
-            }, {
-                label: 'Select All',
-                accelerator: 'Command+A',
-                selector: 'selectAll:'
-            }]
-        }, {
-            label: 'View',
-            submenu: (process.env.NODE_ENV === 'development') ? [{
-                label: 'Reload',
-                accelerator: 'Command+R',
-                click() {
-                    mainWindow.restart();
-                }
-            }, {
-                label: 'Toggle Full Screen',
-                accelerator: 'Ctrl+Command+F',
-                click() {
-                    mainWindow.setFullScreen(!mainWindow.isFullScreen());
-                }
-            }, {
-                label: 'Toggle Developer Tools',
-                accelerator: 'Alt+Command+I',
-                click() {
-                    mainWindow.toggleDevTools();
-                }
-            }] : [{
-                label: 'Toggle Full Screen',
-                accelerator: 'Ctrl+Command+F',
-                click() {
-                    mainWindow.setFullScreen(!mainWindow.isFullScreen());
-                }
-            }]
-        }, {
-            label: 'Window',
-            submenu: [{
-                label: 'Minimize',
-                accelerator: 'Command+M',
-                selector: 'performMiniaturize:'
-            }, {
-                label: 'Close',
-                accelerator: 'Command+W',
-                selector: 'performClose:'
-            }, {
-                type: 'separator'
-            }, {
-                label: 'Bring All to Front',
-                selector: 'arrangeInFront:'
-            }]
-        }, help];
-        menu = Menu.buildFromTemplate(template);
-        Menu.setApplicationMenu(menu);
-    } else {
-        template = [{
-            label: '&File',
-            submenu: [{
-                label: '&Open',
-                accelerator: 'Ctrl+O'
-            }, {
-                label: '&Close',
-                accelerator: 'Ctrl+W',
-                click() {
-                    mainWindow.close();
-                }
-            }]
-        }, {
-            label: '&View',
-            submenu: (process.env.NODE_ENV === 'development') ? [{
-                label: '&Reload',
-                accelerator: 'Ctrl+R',
-                click() {
-                    mainWindow.restart();
-                }
-            }, {
-                label: 'Toggle &Full Screen',
-                accelerator: 'F11',
-                click() {
-                    mainWindow.setFullScreen(!mainWindow.isFullScreen());
-                }
-            }, {
-                label: 'Toggle &Developer Tools',
-                accelerator: 'Alt+Ctrl+I',
-                click() {
-                    mainWindow.toggleDevTools();
-                }
-            }] : [{
-                label: 'Toggle &Full Screen',
-                accelerator: 'F11',
-                click() {
-                    mainWindow.setFullScreen(!mainWindow.isFullScreen());
-                }
-            }]
-        }, help];
-        menu = Menu.buildFromTemplate(template);
-        mainWindow.setMenu(menu);
-    }
 });
